+++ src/app/clusters/actions-server/app_config_dependent_sources.cmake
@@ -18,4 +18,6 @@ TARGET_SOURCES(
   PRIVATE
     "${CLUSTER_DIR}/actions-server.h"
     "${CLUSTER_DIR}/ActionsCluster.cpp"
+    "${CLUSTER_DIR}/ActionsDelegate.h"
+    "${CLUSTER_DIR}/ActionsStructs.h"
 )
