+++ src/app/clusters/actions-server/app_config_dependent_sources.gni
@@ -13,5 +13,7 @@
 # limitations under the License.
 app_config_dependent_sources = [
   "ActionsCluster.cpp",
+  "ActionsDelegate.h",
+  "ActionsStructs.h",
   "actions-server.h",
 ]
