+++ src/app/tests/BUILD.gn
@@ -97,6 +97,8 @@ source_set("power-cluster-test-srcs") {
 source_set("actions-cluster-test-srcs") {
   sources = [
     "${chip_root}/src/app/clusters/actions-server/ActionsCluster.cpp",
+    "${chip_root}/src/app/clusters/actions-server/ActionsDelegate.h",
+    "${chip_root}/src/app/clusters/actions-server/ActionsStructs.h",
     "${chip_root}/src/app/clusters/actions-server/actions-server.h",
   ]
